requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
selenium==4.15.2
webdriver-manager==4.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
stem==1.8.2
PySocks==1.7.1
